feat: Implemented cart and purchase functionalities with UI improvements
- Updated cart UI with improved design and product photo support
- Enhanced product update functionality to include photo changes
- Centered and resized cart page components for better alignment
- Added total price calculations and display for cart items
- Backend improvements for cart item response and order processing
- Minor fixes and optimizations across multiple files

+++ src/main/java/com/hikmetsuicmez/komsu_connect/service/impl/OrderServiceImpl.java
@@ -1,7 +1,6 @@
 package com.hikmetsuicmez.komsu_connect.service.impl;
 
 import com.hikmetsuicmez.komsu_connect.entity.*;
-import com.hikmetsuicmez.komsu_connect.exception.UserNotFoundException;
 import com.hikmetsuicmez.komsu_connect.mapper.OrderItemMapper;
 import com.hikmetsuicmez.komsu_connect.mapper.OrderMapper;
 import com.hikmetsuicmez.komsu_connect.repository.*;
@@ -11,13 +10,16 @@ import com.hikmetsuicmez.komsu_connect.response.OrderResponse;
 import com.hikmetsuicmez.komsu_connect.service.CartService;
 import com.hikmetsuicmez.komsu_connect.service.OrderService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.util.Comparator;
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class OrderServiceImpl implements OrderService {
 
     private final OrderRepository orderRepository;
@@ -29,10 +31,13 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     public OrderResponse createOrder(Long userId) {
+        log.info("Creating order - User ID: {}", userId);
+
         List<CartItemResponse> cartItems = cartService.viewCart(userId);
 
         if (cartItems.isEmpty()) {
-            throw new RuntimeException("No cart items");
+            log.warn("Cart is empty for User ID: {}", userId);
+            throw new RuntimeException("Cart is empty");
         }
 
         double totalPrice = cartItems.stream().mapToDouble(item -> {
@@ -42,35 +47,43 @@ public class OrderServiceImpl implements OrderService {
         }).sum();
 
         User user = userRepository.findById(userId)
-                .orElseThrow(() -> new UserNotFoundException("User Not Found"));
+                .orElseThrow(() -> new RuntimeException("User not found"));
 
         Order order = new Order();
         order.setUser(user);
         order.setTotalPrice(totalPrice);
         order.setCreatedAt(LocalDateTime.now());
         Order savedOrder = orderRepository.save(order);
+        log.info("Order created - Order ID: {}", savedOrder.getId());
 
-        for (CartItemResponse cartItemResponse : cartItems) {
-            Product product = productRepository.findById(cartItemResponse.getProductId())
-                    .orElseThrow(() -> new UserNotFoundException("Product not found"));
-            OrderItem orderItem = OrderItemMapper.toEntity(new OrderItemResponse
-                    (cartItemResponse.getProductId(), cartItemResponse.getProductName(),
-                            cartItemResponse.getQuantity(), product.getPrice()), savedOrder, product);
+        for (CartItemResponse cartItemResponseDTO : cartItems) {
+            Product product = productRepository.findById(cartItemResponseDTO.getProductId())
+                    .orElseThrow(() -> new RuntimeException("Product not found"));
+            OrderItem orderItem = OrderItemMapper.toEntity(new OrderItemResponse(cartItemResponseDTO.getProductId(),
+                    cartItemResponseDTO.getProductName(), cartItemResponseDTO.getQuantity(), product.getPrice()), savedOrder, product);
             orderItemRepository.save(orderItem);
-            CartItem cartItem = cartItemRepository.findById(cartItemResponse.getProductId())
-                    .orElseThrow(() -> new UserNotFoundException("CartItem not found"));
-            cartItemRepository.delete(cartItem);
-        }
+            CartItem cartItem = cartItemRepository.findByProductIdAndUserId(cartItemResponseDTO.getProductId(), userId).orElse(null);
+            if (cartItem != null) {
+                log.info("Deleting CartItem - CartItem ID: {}", cartItem.getId());
+                cartItemRepository.delete(cartItem);
+            } else {
+                log.warn("CartItem not found - Product ID: {}, User ID: {}", cartItemResponseDTO.getProductId(), userId);
+            }
 
+        }
         return OrderMapper.toResponseDto(savedOrder);
     }
 
     @Override
     public List<OrderResponse> getOrderHistory(Long userId) {
+        log.info("Getting order history - User ID: {}", userId);
+
         List<Order> orders = orderRepository.findByUserId(userId);
         return orders
                 .stream()
                 .map(OrderMapper::toResponseDto)
+                .sorted(Comparator.comparing(OrderResponse::getCreatedAt).reversed())
                 .toList();
     }
+
 }
